Lint + setup for tensorboard
